# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Numerical Methods API",
    "version": "1.0.0",
    "description": "API for Numerical Methods"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    }
  ],
  "paths": {
    "/getsample/{path}": {
      "get": {
        "summary": "Get equations by method",
        "description": "Returns equations that match the specified method.",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "description": "The method name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Equation"
                  }
                }
              }
            }
          }
        }
      }
    },
      "/getregression/{path}": {
        "get": {
          "summary": "Get equations by method",
          "description": "Returns equations that match the specified method.",
          "parameters": [
            {
              "name": "path",
              "in": "path",
              "description": "The method name",
              "required": true,
              "schema": {
                "type": "string"
              }
            },{
              "name": "a",
              "in": "query",
              "description": "The parameter a",
              "required": true,
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Equation"
                    }
                  }
                }
              }
            }
          }
        }
      },
    "/gettoken/{path}": {
      "get": {
        "summary": "Get token",
        "description": "Returns a JWT token",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "description": "The method name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuaW1lIjoibWFtYSIsImlhdCI6MTYyMjg1NjA2MCwiZXhwIjoxNjIyODU5NjYwfQ.-QoTSL0ryERHY7vLM4EoHs7sZ1gPXuSCixvYRR6UWV0"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/test/token": {
      "get": {
        "summary": "Test token",
        "description": "Tests if a JWT token is valid",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}